% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Add_HDF5_matrix}
\alias{Add_HDF5_matrix}
\title{Write matrix to existing hdf5 file}
\usage{
Add_HDF5_matrix(mat, filename, group, dataset, transp = NULL)
}
\arguments{
\item{mat}{numerical data matrix}

\item{filename, }{character array indicating the name of the file to create}

\item{group, }{character array indicating folder or group name to put the matrix in hdf5 file}

\item{dataset, }{character array indicating the dataset name that contains the matix data}

\item{transp, }{boolean if true, data is manipulated in transposed form}
}
\value{
none
}
\description{
Creates a hdf5 file with numerical data matrix,
}
