% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdapply_Function_hdf5}
\alias{bdapply_Function_hdf5}
\title{Apply function to different datasets inside a group}
\usage{
bdapply_Function_hdf5(filename, group, datasets, outgroup, func, force = FALSE)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{group, }{character array indicating the input group where the data set to be imputed is.}

\item{datasets, }{character array indicating the input dataset to be imputed}

\item{outgroup, }{optional character array indicating group where the data set will be saved after imputation if `outgroup` is NULL, output dataset is stored in the same input group.}

\item{func, }{character array function to be applyed
\describe{
    \item{QR}{apply bdQR() function to datasets}
    \item{CrossProd}{apply bdCrossprod() function to datasets}
    \item{tCrossProd}{apply bdtCrossprod() function to datasets}
    \item{invChol}{apply bdInvCholesky() function to datasets}
}}

\item{bforce, }{boolean if true, previous results in same location inside hdf5 will be overwritten.}
}
\value{
Original hdf5 data file with results after apply function to different datasets
}
\description{
Apply function to different datasets inside a group
}
