% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdBind_hdf5_datasets}
\alias{bdBind_hdf5_datasets}
\title{Bind matrices by rows or columns}
\usage{
bdBind_hdf5_datasets(
  filename,
  group,
  datasets,
  outgroup,
  outdataset,
  func,
  overwrite = FALSE
)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{group, }{character array indicating the input group where the data set to be imputed is.}

\item{datasets, }{character array indicating the input dataset to be imputed}

\item{outgroup, }{character array indicating group where the data set will be saved after imputation if \code{outgroup} is NULL, output dataset is stored in the same input group.}

\item{outdataset, }{character array indicating the name for the new merged dataset}

\item{func, }{character array function to be applyed
\describe{
\item{bindRows}{merge datasets by rows}
\item{bindCols}{merge datasets by columns}
}}

\item{overwrite, }{boolean if true, previous results in same location inside hdf5 will be overwritten.}
}
\value{
Original hdf5 data file with results after input datasets
}
\description{
Merge existing matrices inside hdf5 data file by rows or by columns
}
