% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdblockmult_hdf5}
\alias{bdblockmult_hdf5}
\title{Hdf5 datasets multiplication}
\usage{
bdblockmult_hdf5(
  filename,
  group,
  A,
  B,
  groupB = NULL,
  block_size = NULL,
  mixblock_size = NULL,
  paral = NULL,
  threads = NULL,
  outgroup = NULL,
  outdataset = NULL,
  force = NULL
)
}
\description{
Multiplies two existing datasets in hdf5 datafile and stores results i a new hdf5 dataset.
}
