% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdblockSubstract_hdf5}
\alias{bdblockSubstract_hdf5}
\title{Hdf5 datasets substract}
\usage{
bdblockSubstract_hdf5(
  filename,
  group,
  A,
  B,
  groupB = NULL,
  block_size = NULL,
  paral = NULL,
  threads = NULL,
  outgroup = NULL,
  outdataset = NULL,
  force = NULL
)
}
\description{
Substracts two existing datasets in hdf5 datafile and stores results i a new hdf5 dataset.
}
