% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdblockMult}
\alias{bdblockMult}
\alias{bdblockSum}
\title{Matrix multiplication}
\usage{
bdblockMult(
  A,
  B,
  block_size = NULL,
  paral = NULL,
  byBlocks = TRUE,
  threads = NULL
)

bdblockSum(
  A,
  B,
  block_size = NULL,
  paral = NULL,
  byBlocks = TRUE,
  threads = NULL
)
}
\arguments{
\item{A}{Matrix A}

\item{B}{Matrix B}

\item{block_size}{PARAM_DESCRIPTION, Default: NULL}

\item{paral}{if paral = TRUE performs parallel computation else performs seria computation, Default: FALSE}

\item{byBlocks}{If data matrix has more than 2.25e+08 (15000 x 15000) elements, by default the addition is done by blocks, but it can be forced not to be partitioned with parameter byblocks = FALSE, Default: TRUE}

\item{threads}{only if bparal = true, number of concurrent threads in parallelization if threads is null then threads =  (maximum number of threads available / 2), Default: NULL}
}
\value{
new matrix with A * B

new matrix with A + B
}
\description{
Miltiplies two existing data matrices on memory

Sum two existing datasets in hdf5 datafile and stores results i a new hdf5 dataset
}
\examples{
\dontrun{
if(interactive()){
    N <- 2500
    M <- 400
    nc <-  4
    
    set.seed(555)
    mat <- matrix( rnorm( N*M, mean=0, sd=10), N, M) 
    
    sum_mem = bdblockMult(mat, mat, paral = TRUE, threads = nc)
 }
}
\dontrun{
if(interactive()){
    N <- 2500
    M <- 400
    nc <-  4
    
    set.seed(555)
    mat <- matrix( rnorm( N*M, mean=0, sd=10), N, M) 
    
    sum_mem = bdblockSum(mat, mat, paral = TRUE, threads = nc)
 }
}
}
