% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocksum_hdf5.R
\name{bdblockSum_hdf5}
\alias{bdblockSum_hdf5}
\title{Block matrix multiplication with hdf5 datasets}
\usage{
bdblockSum_hdf5(
  filename,
  group,
  a,
  b,
  groupB = NULL,
  block_size = 128,
  paral = FALSE,
  threads = NULL,
  outgroup = "OUTPUT",
  outdataset = NULL
)
}
\arguments{
\item{filename}{string file name where dataset to normalize is stored}

\item{group}{string with the group name where matrix is stored inside HDF5 file}

\item{a}{a double matrix.}

\item{b}{a double matrix.}

\item{groupB, }{string, (optional) group name where dataset b is stored}

\item{block_size}{(optional, defalut = 128) block size to make matrix multiplication, if `block_size = 1` no block size is applied (size 1 = 1 element per block)}

\item{paral, }{(optional, default = FALSE) paral = true --> TO BE IMPLEMENTED}

\item{threads}{(optional) only if bparal = true, number of concurrent threads in parallelization if threads is null then threads =  maximum number of threads available}

\item{outgroup}{(optional) string with group name where we want to store the result matrix}

\item{outdataset}{(optional) string with dataset name where we want to store the results}

\item{mixblock_size}{(optiona, default = 128), only if we are working with big matrix and parallel computation = true.
Block size for mixed computation in big matrix parallel. Size of the block to be used to perform parallelized memory
memory of the block read from the disk being processed.}
}
\value{
A list with an HDF5 object with numerical matrix and  HDF5 file name with results
\itemize{
 #### \item{"res"}{rhdf5 object with result matrix - link to hdf5 file contents . IMPORTANT !!, we have to close the object after fihish to work}
 \item{"file"}{string with hdf5 file name with result and original matrixs}
 \item{"dataset"}{string complete path inside hdf5 file where results are stored.}
}
}
\description{
This function performs a block matrix-matrix multiplication with mattrix stored in HDF5 file
}
\examples{
# with numeric matrix


m <- 500
k <- 1500
n <- 400
A <- matrix(rnorm(n*k), nrow=n, ncol=k)
B <- matrix(rnorm(n*k), nrow=k, ncol=n)

C <- bdblockSum_hdf5(A, B, 128, TRUE)

}
