% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdcloseres.R
\name{bdclose}
\alias{bdclose}
\title{Close files from hdf5 objects}
\usage{
bdclose(results)
}
\arguments{
\item{results}{object from on disk execution}
}
\value{
void
}
\description{
This function closes the files associated with the hdf5 objects of the operations performed on disk.
}
\examples{
# with numeric matrix

m <- 350
n <- 150
k <- 50

Abig <- matrix(rnorm(m*k), nrow=m, ncol=k)
Bbig <- matrix(rnorm(k*n), nrow=k, ncol=n)

DAbig <- DelayedArray(Abig)
DBbig <- DelayedArray(Bbig)

# We consider a big matrix if number of rows or columns are > 500
C <- blockmult(DAbig, DBbig, bigmatrix = 1, paral = FALSE)

bdclose(C)
}
