% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdCreateGroup_hdf5}
\alias{bdCreateGroup_hdf5}
\title{Create groups}
\usage{
bdCreateGroup_hdf5(filename, group)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{group, }{string with the name for the new group (complete route)}
}
\value{
none
}
\description{
Create groups in hdf5 data file
}
\examples{

matA <- matrix(c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15), nrow = 3, byrow = TRUE)

bdCreate_hdf5_matrix_file("BasicMatVect.hdf5", matA, "INPUT", "matA", force = TRUE)
bdCreateGroup_hdf5("BasicMatVect.hdf5", "INPUT/NEWGROUP")
bdCreateGroup_hdf5("BasicMatVect.hdf5", "NEWGROUP2")

# Remove file (used as example)
  if (file.exists("BasicMatVect.hdf5")) {
    # Delete file if it exist
    file.remove("BasicMatVect.hdf5")
  }

}
