% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdCreateLink_hdf5}
\alias{bdCreateLink_hdf5}
\title{Create hard link between two datasets}
\usage{
bdCreateLink_hdf5(filename, source, dest)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{source, }{string with route to source dataset}

\item{dest, }{string with route to destination dataset}
}
\value{
none
}
\description{
Create hard link between two datasets
}
