% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdExists_hdf5_element}
\alias{bdExists_hdf5_element}
\title{Exists hdf5 element}
\usage{
bdExists_hdf5_element(filename, element)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{element, }{string with name of the group or dataset full route group or dataset to search}
}
\value{
boolean, true if element exists in hdf5 data faile or false if not.
}
\description{
Query if exists element inside hdf5 data file
}
\examples{

# Prepare data to write dataset inside a file and test if exists
X <- matrix(rnorm(20*10), nrow = 10, ncol = 20)

# Create hdf5 data file with  two X matrices ( named X and Y)
bdCreate_hdf5_matrix_file("test_file3.hdf5", X, "data", "X")
bdAdd_hdf5_matrix(X, "test_file3.hdf5", "data", "Y")

bdExists_hdf5_element("test_file3.hdf5", "data/X")
bdExists_hdf5_element("test_file3.hdf5", "data/F")

}
