% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdgetDiagonal_hdf5}
\alias{bdgetDiagonal_hdf5}
\title{Write diagonal matrix}
\usage{
bdgetDiagonal_hdf5(filename, group, dataset)
}
\arguments{
\item{filename, }{character array with the name of an existin hdf5 data file containing the dataset to be modified}

\item{group, }{character array indicating the input group where the data set to be modified.}

\item{datasets, }{character array indicating the input dataset to be modified}
}
\value{
Original hdf5 dataset with new diagonal elements
}
\description{
Write diagonal matrix to an existing dataset inside hdf5
}
