% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdMLR_MR}
\alias{bdMLR_MR}
\title{Linear regression using MLR-MR algorithm}
\usage{
bdMLR_MR(X, y, blocks, threads = NULL)
}
\arguments{
\item{X, }{numerical matrix with paired observations of the predictor variable X}

\item{blocks, }{integer with number of blocks we want to split matrix if null matrix is splited in blocks as maximum of 1000 variables per block}

\item{threads, }{threads (optional) only if bparal = true, number of concurrent threads in parallelization if threads is null then threads =  maximum number of threads available}

\item{Y, }{numerical matrix column with response variable}
}
\value{
coefficients and error
}
\description{
Linear regression for Big Data using MLR-MR algorithm
}
