% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdNormalize_hdf5}
\alias{bdNormalize_hdf5}
\title{Normalize dataset in hdf5 file}
\usage{
bdNormalize_hdf5(
  filename,
  group,
  dataset,
  bcenter = NULL,
  bscale = NULL,
  byrows = NULL,
  wsize = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{filename}{string file name where dataset to normalize is stored}

\item{group}{string specifying the group within the HDF5 file containing
matrix dataset.}

\item{dataset}{string, a string specifying the name of the dataset to
perform calculus.}

\item{bcenter}{logical (optional). If TRUE (default), the data is centered by
subtracting the column means (ignoring NAs) of \code{x} from their corresponding columns.
If FALSE, no centering is performed.}

\item{bscale}{(optional). If TRUE (default), the data is scaled by dividing
the (centered) columns of \code{x} by their standard deviations if \code{bcenter} is TRUE,
or by the root mean square otherwise. If FALSE, no scaling is performed.}

\item{byrows}{logical (default = FALSE) if TRUE, centering is done by
subtracting the rows means, util when working with hdf5 datasets stored
in Row Major format.}

\item{wsize}{integer (default = 1000), file block size to read to
perform normalization}

\item{overwrite, }{boolean if true, previous results in same location inside
hdf5 will be overwritten.}
}
\value{
the original HDF5 file with normalized data stored under the group
"NORMALIZED", where:
* the dataset for the mean is named "mean." + original_dataset_name
* the dataset for the scaling is named "sd." + original_dataset_name
}
\description{
This function normalize data scaling, centering or scaling and centering
in a dataset stored in hdf5 file
}
\examples{
  a = "See vignette"
}
