% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdparallelpow2}
\alias{bdparallelpow2}
\title{Pow vector}
\usage{
bdparallelpow2(x)
}
\arguments{
\item{x}{numerical vector}
}
\description{
Gets pow2 vector
}
\examples{
library(BigDataStatMeth)

n <- 100 
x <- rnorm(n)

# with numeric matrix
res <- bdparallelpow2(x)

}
