% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdparallelVectorSum}
\alias{bdparallelVectorSum}
\title{Sumarize vector}
\usage{
bdparallelVectorSum(x)
}
\arguments{
\item{x}{numerical vector}
}
\description{
This function sumarize the elements of a vector
}
\examples{
library(BigDataStatMeth)

n <- 100 
x <- rnorm(n)

# with numeric matrix
res <- bdparallelVectorSum(x)

}
