% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdQR_compact}
\alias{bdQR_compact}
\title{QR Decomposition}
\usage{
bdQR_compact(A)
}
\arguments{
\item{X}{a real square matrix}
}
\value{
List with  : a matrix with the same dimensions as x. The upper triangle contains the \(\bold{R}\) of the decomposition and the lower triangle contains information on the \(\bold{Q}\) of the decomposition (stored in compact form)
}
\description{
This function compute QR decomposition computes a QR
factorization with column pivoting of a matrix A:  A*P = Q*R
}
\details{
This QR decomposition is like R native decomposition obtained with qr() function
}
