% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdQR_parallel}
\alias{bdQR_parallel}
\title{QR Decomposition}
\usage{
bdQR_parallel(X, threads = NULL)
}
\arguments{
\item{X}{a real square matrix}

\item{threads}{(optional) only if paral = true, number of concurrent threads in parallelization if threads is null then threads =  maximum number of threads available}
}
\value{
List with orthogonal matrix \code{Q}  and upper triangular matrix \code{R}
}
\description{
This function compute QR decomposition (also called a QR factorization) 
of a matrix \code{A} into a product \code{A = QR} of an 
orthogonal matrix Q and an upper triangular matrix R.
}
