% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdReduce_matrix_hdf5}
\alias{bdReduce_matrix_hdf5}
\title{Reduce hdf5 dataset}
\usage{
bdReduce_matrix_hdf5(
  filename,
  group,
  reducefunction,
  outgroup = NULL,
  outdataset = NULL,
  force = FALSE,
  remove = FALSE
)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{group, }{character array indicating the input group where the data sets are stored}

\item{reducefunction, }{single character with function to apply, can be '+' or  '-' array indicating the input dataset to be imputed}

\item{outgroup, }{optional character array indicating group where the data set will be saved after imputation if `outgroup` is NULL, output dataset is stored in the same input group.}

\item{outdataset, }{optional character array indicating dataset to store the resulting data after imputation if `outdataset` is NULL, input dataset will be overwritten.}

\item{force, }{boolean if true, previous results in same location inside hdf5 will be overwritten.}

\item{remove, }{boolean if true, removes original matrices, by default bremove = false.}
}
\value{
Full matrix with results from reduction
}
\description{
Reduce hdf5 datasets inside a group by rows or columns and store complete matrix inside hdf5 data file.
}
