% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BDsvd}
\alias{BDsvd}
\title{SVD of DelayedArray}
\usage{
BDsvd(x, k = 0L, nev = 0L, normalize = TRUE)
}
\arguments{
\item{x}{An integer vector}

\item{k}{....}

\item{nev}{...}

\item{normalize}{...}
}
\description{
This function ....
}
