% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdWeightedProduct_hdf5}
\alias{bdWeightedProduct_hdf5}
\title{Apply weights to a dataset in hdf5 file}
\usage{
bdWeightedProduct_hdf5(
  filename,
  group,
  dataset,
  vectorgroup,
  vectordataset,
  outdataset,
  outgroup = NULL,
  byrows = NULL,
  force = FALSE
)
}
\arguments{
\item{filename}{string file name where dataset to apply weights is located}

\item{group}{string with the path inside the hdf5 data file where matrix 
is located}

\item{dataset}{string with the matrix name}

\item{vectorgroup}{string with the path inside the hdf5 data file where 
vector is located}

\item{vectordataset}{with the vector name}

\item{outdataset}{character array with output dataset name where we want to 
store results}

\item{outgroup}{optional, character array with output group name where we 
want to store results if not provided then results are stored in the same 
group as original dataset}

\item{byrows}{logical (default = FALSE). By default weights are applied by 
columns but if byrows=TRUE then weights are applied by rows}

\item{force, }{boolean if true, previous results in same location inside 
hdf5 will be overwritten.}
}
\value{
file with weighted dataset
}
\description{
This function applies a weight vector to a matrix. Multiplies each matrix 
row/column from a hdf5 dataset with a weight vector
}
\examples{
library(BigDataStatMeth)
   
# Prepare data and functions
set.seed(123)
Y <- matrix(rnorm(250), 5, 10)
X <- matrix(rnorm(250), 10, 1)
       
# Create hdf5 data file with  data (Y)
bdCreate_hdf5_matrix_file("cca_cars.hdf5", Y, "data", "Y", force = T)
bdAdd_hdf5_matrix( X, "cca_cars.hdf5",  "data", "X", force = TRUE)
           
bdWeightedProduct_hdf5("cca_cars.hdf5", 
                      group = "data", dataset = "Y",
                      vectorgroup = "data", vectordataset = "X", 
                      outdataset = "Prod", byrows = T, force = T)
}
