% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdWriteDimnames_hdf5}
\alias{bdWriteDimnames_hdf5}
\title{Write dimnames in HDF5}
\usage{
bdWriteDimnames_hdf5(filename, group, dataset, rownames, colnames)
}
\arguments{
\item{filename, }{character array indicating the name of the file}

\item{group, }{string with name of the group where the dataset is located}

\item{dataset, }{string with the dataset name}

\item{rownames, }{character vector, with the rownames, if rownames is NULL 
no rownames are written to the dataset}

\item{colnames, }{character vector, with the colnames, if colnames is NULL
no colnames are written to the dataset}
}
\value{
none
}
\description{
Write dimnames in HDF5 related to an existing dataset
}
\examples{

# Writes dimnames

}
