% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdwXw}
\alias{bdwXw}
\title{Matrix - Weighted vector Multiplication with numerical or DelayedArray data}
\usage{
bdwXw(X, W, op)
}
\arguments{
\item{X}{numerical or Delayed Array matrix}

\item{W}{vector with weights}

\item{op}{string indicating if operation  "Xw" , "wX" or "wXw"}
}
\value{
numerical matrix
}
\description{
This function performs a weighted product of a matrix(X) with a weighted diagonal matrix (w)
}
\examples{

library(DelayedArray)

n <- 100
p <- 60

X <- matrix(rnorm(n*p), nrow=n, ncol=p)

u <- runif(n)
w <- u * (1 - u)

# by columnes
bdwXw(X, w,"wX") 

# with Delayed Array

DX <- DelayedArray(X)

bdwXw(DX, w,"wX")

}
