% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blockCrossprod_hdf5}
\alias{blockCrossprod_hdf5}
\title{Crossprod with hdf5 matrix}
\usage{
blockCrossprod_hdf5(
  filename,
  group,
  A,
  block_size = NULL,
  paral = NULL,
  threads = NULL,
  mixblock_size = NULL,
  outgroup = NULL
)
}
\arguments{
\item{filename}{string file name where dataset to normalize is stored}

\item{group}{string or Delayed Array Matrix}

\item{block_size}{(optional, defalut = 128) block size to make matrix multiplication, if `block_size = 1` no block size is applied (size 1 = 1 element per block)}

\item{paral, }{(optional, default = TRUE) if paral = TRUE performs parallel computation else performs seria computation}

\item{threads}{(optional) only if bparal = true, number of concurrent threads in parallelization if threads is null then threads =  maximum number of threads available}

\item{outgroup}{(optional) group name to store results from Crossprod inside hdf5 data file}

\item{dataset}{string name inside HDF5 file}
}
\description{
This function performs the crossprod from a matrix inside and hdf5 data file
}
\examples{
  a = "See vignette"
}
