% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blockmult_hdf5}
\alias{blockmult_hdf5}
\title{Multiply hdf5 matrix}
\usage{
blockmult_hdf5(
  filename,
  group,
  A,
  B,
  block_size = NULL,
  paral = NULL,
  threads = NULL,
  bigmatrix = NULL,
  mixblock_size = NULL,
  outfile = NULL,
  outgroup = NULL
)
}
\arguments{
\item{filename}{string file name where dataset to normalize is stored}

\item{group}{string or Delayed Array Matrix}

\item{dataset}{string or Delayed Array Matrix}

\item{bcenter}{logical (default = TRUE) if TRUE, centering is done by subtracting the column means}

\item{bscale}{logical (default = TRUE) if TRUE, centering is done by subtracting the column means}

\item{wsize}{integer (default = 1000), file block size to read to perform normalization}
}
\value{
file with scaled, centered or scaled and centered dataset
}
\description{
This function multiply matrix stored in hdf5 data file
}
\examples{
  a = "See vignette"
}
