% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Cancer classification}
\format{
factor level with three levels
\describe{
\item{cancer}{factor with cancer type}
}
}
\usage{
data(cancer)
}
\description{
A three factor level variable corresponding to cancer type
}
\examples{
data(cancer)
}
\keyword{datasets}
