% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{colesterol}
\alias{colesterol}
\title{Dataset colesterol}
\description{
This dataset contains a dummy data for import dataset example
}
\section{colesterol.csv}{


This data is used in bdImportTextFile_hdf5() function.
}

