% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Create_HDF5_matrix}
\alias{Create_HDF5_matrix}
\title{Write matrix to existing hdf5 file}
\usage{
Create_HDF5_matrix(mat, filename, group, dataset, transp = NULL)
}
\arguments{
\item{mat}{numerical data matrix}

\item{filename, }{character array indicating the name of the file to create}

\item{dataset, }{character array indicating the dataset name that contains the matix data}

\item{folder, }{character array indicating folder name to put the matrix in hdf5 file}
}
\value{
none
}
\description{
Creates a hdf5 file with numerical data matrix,
}
