% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Create_HDF5_matrix_file}
\alias{Create_HDF5_matrix_file}
\title{Write data to hdf5 file}
\usage{
Create_HDF5_matrix_file(filename, mat, folder = NULL)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{mat}{numerical data matrix}

\item{folder, }{character array indicating folder name to put the matrix in hdf5 file}
}
\value{
none
}
\description{
Creates a hdf5 file with numerical data matrix,
}
