% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_dimnames.R
\name{dimnames<-.HDF5Matrix}
\alias{dimnames<-.HDF5Matrix}
\title{Set dimension names on an HDF5Matrix}
\usage{
\method{dimnames}{HDF5Matrix}(x) <- value
}
\arguments{
\item{x}{An \code{HDF5Matrix} object}

\item{value}{A list of length 2: \code{list(rownames, colnames)}.
Either element may be \code{NULL} to skip that dimension.}
}
\value{
\code{x} invisibly
}
\description{
Writes row and/or column names to the HDF5 file alongside the dataset.
Setting a dimension name to \code{NULL} removes names for that dimension.
}
