% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{.hdf5matrix_options}
\alias{.hdf5matrix_options}
\title{HDF5Matrix Global Options}
\format{
An object of class \code{environment} of length 4.
}
\usage{
.hdf5matrix_options
}
\description{
Internal environment to store global computation options for HDF5Matrix operations.
}
\keyword{internal}
