% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5CloseAll.R
\name{hdf5_list_open}
\alias{hdf5_list_open}
\title{List open HDF5 files}
\usage{
hdf5_list_open()
}
\value{
Character vector of open file paths
}
\description{
List all currently open HDF5 files in the R session.
Useful for debugging file locking issues.
}
\examples{
\donttest{
tmp <- tempfile(fileext = ".h5")
rhdf5::h5createFile(tmp)
rhdf5::h5write(matrix(1:100, 10, 10), tmp, "data/matrix")

X <- hdf5_matrix(tmp, "data/matrix")

# Check open files
hdf5_list_open()  # Shows tmp file

X$close()
hdf5_list_open()  # Empty

unlink(tmp)
}

}
