% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_algebra.R
\name{\%*\%}
\alias{\%*\%}
\title{Matrix multiplication for HDF5Matrix}
\usage{
x \%*\% y
}
\arguments{
\item{x}{Left-hand side matrix.}

\item{y}{Right-hand side matrix.}
}
\value{
Result matrix.
}
\description{
S3 generic for \code{\%*\%}. Dispatches to \code{\%*\%.HDF5Matrix}
for \code{HDF5Matrix} objects, and to \code{base::\%*\%} for all others.
}
