% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNA}
\alias{miRNA}
\title{miRNA}
\format{
Dataframe with 21 samples and 537 variables
\describe{
\item{columns}{variables}
\item{rows}{samples}
}
}
\usage{
data(miRNA)
}
\description{
A three factor level variable corresponding to cancer type
}
\examples{
data(miRNA)
}
\keyword{datasets}
