% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_mr.R
\name{mlr_mr}
\alias{mlr_mr}
\title{Linear regression using MLR-MR algorithm}
\usage{
mlr_mr(Y, model, blocks, threads = NULL)
}
\arguments{
\item{Y, }{numerical matrix column with response variable}

\item{blocks, }{integer with number of blocks we want to split matrix if null matrix is splited in blocks as maximum of 1000 variables per block}

\item{threads, }{threads (optional) only if bparal = true, number of concurrent threads in parallelization if threads is null then threads =  maximum number of threads available}

\item{X, }{numerical matrix with paired observations of the predictor variable X}
}
\value{
coefficients
}
\description{
Linear regression for Big Data using MLR-MR algorithm
}
\examples{
# with numeric matrix

library(BigDataStatMeth)
data(mtcars)

Y <- mtcars$mpg
X <- model.matrix(~ wt + cyl, data=mtcars)
m <- 7


res <- bdMLR_MR( X, Y, m, 1)
res

}
