% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Normalize_Data}
\alias{Normalize_Data}
\title{Normalize Delayed Array matrix}
\usage{
Normalize_Data(x)
}
\arguments{
\item{X}{numerical or Delayed Array Matrix}
}
\value{
numerical matrix
}
\description{
This function performs a numerical or Delayed Array matrix normalization
}
\examples{
m <- 500
n <- 100 
x <- matrix(rnorm(m*n), nrow=m, ncol=n)

# with numeric matrix
Normalize_Data(x)

# with Delaeyd Array
Dx <- DelayedArray(x)
Normalize_Data(Dx)

}
