% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallelpow2}
\alias{parallelpow2}
\title{Pow vector}
\usage{
parallelpow2(x)
}
\arguments{
\item{x}{numerical vector}
}
\description{
Gets pow2 vector
}
\examples{
library(BigDataStatMeth)

n <- 100 
x <- rnorm(n)

# with numeric matrix
parallelpow2(x)

}
