% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parXYProd}
\alias{parXYProd}
\title{Matrix multiplication with Delayed Array Object (RcppParallel)}
\usage{
parXYProd(X, Y, op = NULL)
}
\arguments{
\item{X}{double matrix to apply multiplication}

\item{Y}{double matrix to apply multiplication}

\item{op, }{(optional, default = "xy"), if op="xy" then performs the x\%*\%y matrix multiplication, if op = "xty" preforms t(X)\%*\% Y, if op = "xyt" performs X\%*\%t(Y)}
}
\value{
numerical matrix
}
\description{
This function performs a block matrix-matrix multiplication with numeric matrix or Delayed Arrays
}
\examples{

library(DelayedArray)

# with numeric matrix
k <- 300
n <- 400
A <- matrix(rnorm(n*k), nrow=n, ncol=k)
B <- matrix(rnorm(k*k), nrow=k)

parXYProd(A,B,"xy")
parXYProd(A,B,"xyt")

# with Delaeyd Array
AD <- DelayedArray(A)
BD <- DelayedArray(B)

parXYProd(AD,BD,"xy")
parXYProd(AD,BD,"xyt")

}
