% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parXYProdBlock}
\alias{parXYProdBlock}
\title{Block matrix multiplication with Delayed Array Object}
\usage{
parXYProdBlock(X, Y, op = NULL)
}
\arguments{
\item{X}{double matrix to apply multiplication}

\item{Y}{double matrix to apply multiplication}

\item{op, }{(optional, default = "xy"), if op="xy" then performs the x\%*\%y matrix multiplication, if op = "xty" preforms t(X)\%*\% Y, if op = "xyt" performs X\%*\%t(Y)}
}
\value{
numerical matrix
}
\description{
This function performs a block matrix-matrix multiplication with numeric matrix or Delayed Arrays
}
\examples{

library(DelayedArray)

# with numeric matrix
m <- 500
k <- 1500
n <- 400
A <- matrix(rnorm(n*k), nrow=n, ncol=k)
B <- matrix(rnorm(n*k), nrow=k, ncol=n)

parXYProdBlock(A,B,"xy")
parXYProdBlock(A,B,"xty")

# with Delaeyd Array
AD <- DelayedArray(A)
BD <- DelayedArray(B)

parXYProdBlock(AD,BD,"xy")
parXYProdBlock(AD,BD,"xty")


}
