% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_core.R
\name{print.HDF5Matrix}
\alias{print.HDF5Matrix}
\title{Print an HDF5Matrix object}
\usage{
\method{print}{HDF5Matrix}(x, ...)
}
\arguments{
\item{x}{An \code{HDF5Matrix} object}

\item{...}{Ignored}
}
\value{
Invisible \code{x}
}
\description{
Print an HDF5Matrix object
}
\examples{
\donttest{
tmp <- tempfile(fileext = ".h5")
rhdf5::h5createFile(tmp)
rhdf5::h5createGroup(tmp, "data")
rhdf5::h5write(matrix(1:100, 10, 10), tmp, "data/matrix")
X <- hdf5_matrix(tmp, "data/matrix")

print(X)
X   # same as print(X)

X$close()
unlink(tmp)
}

}
