% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_decompositions.R
\name{print.HDF5PCA}
\alias{print.HDF5PCA}
\title{Print method for HDF5PCA objects}
\usage{
\method{print}{HDF5PCA}(x, ...)
}
\arguments{
\item{x}{An \code{HDF5PCA} object returned by \code{prcomp()} or \code{$pca()}.}

\item{...}{Ignored.}
}
\description{
Print method for HDF5PCA objects
}
