% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hdf5_close_file_handles_safe}
\alias{rcpp_hdf5_close_file_handles_safe}
\title{Safely close all remaining HDF5 file handles (mid-session safe)}
\usage{
rcpp_hdf5_close_file_handles_safe()
}
\description{
Safely close all remaining HDF5 file handles (mid-session safe)
}
\keyword{internal}
