% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hdf5dataset_dim}
\alias{rcpp_hdf5dataset_dim}
\title{Get dimensions of HDF5 dataset (R6 wrapper)}
\usage{
rcpp_hdf5dataset_dim(ptr_sexp)
}
\arguments{
\item{ptr_sexp}{External pointer to hdf5Dataset}
}
\value{
Integer vector c(nrows, ncols)
}
\description{
Get dimensions of HDF5 dataset (R6 wrapper)
}
\keyword{internal}
