% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hdf5dataset_info}
\alias{rcpp_hdf5dataset_info}
\title{Get dataset information (R6 wrapper)}
\usage{
rcpp_hdf5dataset_info(ptr_sexp)
}
\arguments{
\item{ptr_sexp}{External pointer to hdf5Dataset}
}
\value{
Named list with filename, group, dataset, datatype
}
\description{
Get dataset information (R6 wrapper)
}
\keyword{internal}
