% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hdf5dataset_is_valid}
\alias{rcpp_hdf5dataset_is_valid}
\title{Check if dataset is valid and open (R6 wrapper)}
\usage{
rcpp_hdf5dataset_is_valid(ptr_sexp)
}
\arguments{
\item{ptr_sexp}{External pointer to hdf5Dataset}
}
\value{
Logical: TRUE if valid and open, FALSE otherwise
}
\description{
Check if dataset is valid and open (R6 wrapper)
}
\keyword{internal}
