% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hdf5dataset_read_all}
\alias{rcpp_hdf5dataset_read_all}
\title{Get full dataset as matrix (convenience function)}
\usage{
rcpp_hdf5dataset_read_all(ptr_sexp)
}
\arguments{
\item{ptr_sexp}{External pointer to hdf5Dataset}
}
\value{
Numeric matrix with all data
}
\description{
Get full dataset as matrix (convenience function)
}
\keyword{internal}
