% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hdf5dataset_write_all}
\alias{rcpp_hdf5dataset_write_all}
\title{Write entire dataset (R6 wrapper)}
\usage{
rcpp_hdf5dataset_write_all(ptr_sexp, value)
}
\arguments{
\item{ptr_sexp}{External pointer (SEXP) to hdf5Dataset}

\item{value}{Data to write (numeric matrix)}
}
\value{
NULL (invisible)
}
\description{
Replaces entire HDF5 dataset contents with new data.
}
\keyword{internal}
