% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_operations.R
\name{reduce}
\alias{reduce}
\title{Reduce a group of HDF5 datasets by accumulation (generic)}
\usage{
reduce(x, ...)
}
\arguments{
\item{x}{An \code{HDF5Matrix}.}

\item{...}{Additional arguments forwarded to the method.}
}
\value{
A new \code{HDF5Matrix} containing the accumulated result.
}
\description{
Generic function for reducing (accumulating) all datasets in the same HDF5
group as \code{x} into a single dataset using a binary operation.
}
\examples{
\donttest{
partial <- hdf5_matrix("data.h5", "partials/chunk_0")
total   <- reduce(partial, func = "+")   # sums all datasets in "partials/"
}

}
\seealso{
\code{\link{reduce.HDF5Matrix}}
}
