% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Remove_HDF5_element}
\alias{Remove_HDF5_element}
\title{Remove element group or dataset from  hdf5 file}
\usage{
Remove_HDF5_element(filename, element)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{element}{path to element, character array indicating the complete route to the element to be removed (folder or dataset).}
}
\value{
none
}
\description{
Remove group or dataset from  hdf5 file
}
