% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{show_hdf5matrix_options}
\alias{show_hdf5matrix_options}
\title{Show current HDF5Matrix performance settings}
\usage{
show_hdf5matrix_options()
}
\value{
Invisibly returns the options list
}
\description{
Display current global options in a user-friendly format.
}
\examples{
show_hdf5matrix_options()

}
