% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_core.R
\name{str.HDF5Matrix}
\alias{str.HDF5Matrix}
\title{Structure of an HDF5Matrix object}
\usage{
\method{str}{HDF5Matrix}(object, ...)
}
\arguments{
\item{object}{An \code{HDF5Matrix} object}

\item{...}{Ignored}
}
\value{
Invisible \code{object}
}
\description{
Structure of an HDF5Matrix object
}
\examples{
\donttest{
tmp <- tempfile(fileext = ".h5")

X  <- hdf5_create_matrix(tmp, "data/matrix", data = matrix(rnorm(100), 10, 10))
X <- hdf5_matrix(tmp, "data/matrix")

str(X)

X$close()
unlink(tmp)
}

}
