% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{system_info}
\alias{system_info}
\title{Get system information summary}
\usage{
system_info()
}
\value{
Named list with system information:
\describe{
\item{os}{Operating system name}
\item{total_ram_gb}{Total RAM in GB}
\item{available_ram_gb}{Available RAM in GB}
\item{ram_used_pct}{Percentage of RAM currently used}
\item{cpu_cores}{Number of CPU cores}
}
}
\description{
Returns a comprehensive summary of system resources.
}
\details{
Convenience function that calls all system info methods and
returns a summary. Useful for debugging and logging.
}
\examples{
\donttest{
# Get full system info
info <- system_info()
print(info)
# $os: "Linux"
# $total_ram_gb: 16.0
# $available_ram_gb: 8.2
# $ram_used_pct: 48.75
# $cpu_cores: 8
}

}
