% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tCrossprod_Weighted}
\alias{tCrossprod_Weighted}
\title{Block matrix multiplication with Delayed Array Object}
\usage{
tCrossprod_Weighted(a, w, block_size = NULL, paral = NULL, threads = NULL)
}
\arguments{
\item{a}{a double matrix.}

\item{w}{a Weighted matrix}

\item{block_size}{(optional, defalut = 128) block size to make matrix multiplication, if `block_size = 1` no block size is applied (size 1 = 1 element per block)}

\item{paral, }{(optional, default = TRUE) if paral = TRUE performs parallel computation else performs seria computation}

\item{threads}{(optional) only if bparal = true, number of concurrent threads in parallelization if threads is null then threads =  maximum number of threads available}
}
\value{
Matrix with crossproduct :
}
\description{
This function performs a Crossproduct with weigths matrix A%*%W%*%t(A) multiplication with numeric matrix or Delayed Arrays
}
\examples{

library(DelayedArray)

# with numeric matrix
m <- 500
k <- 1500
n <- 400
A <- matrix(rnorm(n*k), nrow=n, ncol=k)
B <- matrix(rnorm(n*k), nrow=k, ncol=n)

blockmult(A,B,128, TRUE)

# with Delaeyd Array
AD <- DelayedArray(A)
BD <- DelayedArray(B)

blockmult(AD,BD,128, TRUE)
}
