% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports 2.R, R/RcppExports.R
\name{bdblockSum_hdf5}
\alias{bdblockSum_hdf5}
\title{Hdf5 datasets sum}
\usage{
bdblockSum_hdf5(
  filename,
  group,
  A,
  B,
  groupB = NULL,
  block_size = NULL,
  paral = NULL,
  threads = NULL,
  outgroup = NULL,
  outdataset = NULL,
  force = NULL
)

bdblockSum_hdf5(
  filename,
  group,
  A,
  B,
  groupB = NULL,
  block_size = NULL,
  paral = NULL,
  threads = NULL,
  outgroup = NULL,
  outdataset = NULL,
  force = NULL
)
}
\description{
Sum two existing datasets in hdf5 datafile and stores results i a new hdf5 dataset.

Sum two existing datasets in hdf5 datafile and stores results i a new hdf5 dataset.
}
