% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bddtrsm}
\alias{bddtrsm}
\title{Solves matrix equations : A*X = B}
\usage{
bddtrsm(R, Z, threads = NULL)
}
\arguments{
\item{R}{numerical or Delayed Array matrix.}

\item{Z}{numerical or Delayed Array matrix.}
}
\value{
X numerical matrix.
}
\description{
Solves matrix equations : A*X = B
}
\examples{

}
