% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdgetDatasetsList_hdf5}
\alias{bdgetDatasetsList_hdf5}
\title{Reduce hdf5 dataset}
\usage{
bdgetDatasetsList_hdf5(filename, group, prefix = NULL)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{group, }{character array indicating the input group where the data sets are stored}

\item{prefix, }{character array optional, indicates the prefix with which the dataset
names begin, if null, then the function returns all datasets inside the group}
}
\value{
Full matrix with results from reduction
}
\description{
Reduce hdf5 datasets inside a group by rows or columns and store complete matrix inside hdf5 data file.
}
