% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdgetDiagonal_hdf5}
\alias{bdgetDiagonal_hdf5}
\title{Get diagonal matrix}
\usage{
bdgetDiagonal_hdf5(filename, group, dataset)
}
\arguments{
\item{filename, }{character array with the name of an existin hdf5 data file containing the dataset to be modified}

\item{group, }{character array indicating the input group where dataset is stored}

\item{dataset, }{character array indicating the input dataset name}
}
\value{
Numeric vector with all diagonal elements from hdf5 dataset
}
\description{
Gry diagonal matrix from an existing dataset inside hdf5
}
\examples{

library(BigDataStatMeth)
library(rhdf5)

X <- matrix(rnorm(100), 10, 10)
diag(X) <- 0.5
# Create hdf5 data file with  data (Y)
bdCreate_hdf5_matrix_file("test_file2.hdf5", X, "data", "X", force = TRUE)
# Update diagonal
diagonal <- bdgetDiagonal_hdf5("test_file.hdf5", "data", "X")

}
