% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdgetDim_hdf5}
\alias{bdgetDim_hdf5}
\title{Get dataset dimensions}
\usage{
bdgetDim_hdf5(filename, element)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{element}{path to element, character array indicating the complete route to the element to query size (folder or dataset).}
}
\value{
none
}
\description{
get dataset dimensions
}
