% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdgetDim_hdf5}
\alias{bdgetDim_hdf5}
\title{Get hdf5 dataset dimensions}
\usage{
bdgetDim_hdf5(filename, dataset)
}
\arguments{
\item{filename, }{character array with the name of the file where datasets are stored}

\item{dataset, }{character array with the input dataset name, with complete route inside the hdf5 file}
}
\value{
Dataset dimensions
}
\description{
Get hdf5 dataset dimensions
}
