% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdgetSDandMean_hdf5}
\alias{bdgetSDandMean_hdf5}
\title{Get sd and Mean by Rows or Columns}
\usage{
bdgetSDandMean_hdf5(
  filename,
  group,
  dataset,
  sd = NULL,
  mean = NULL,
  byrows = NULL,
  wsize = NULL,
  force = FALSE
)
}
\arguments{
\item{filename}{string file name where dataset to normalize is stored}

\item{group}{string Matrix}

\item{dataset}{string Matrix}

\item{sd}{logical (default = TRUE) if TRUE, standard deviation is computed}

\item{mean}{logical (default = TRUE) if TRUE, mean is computed}

\item{byrows}{logical (default = FALSE) if TRUE, sd and mean are computed
by columns, if byrows=TRUE then sd and mean are computed by Rows.}

\item{wsize}{integer (default = 1000), file block size to read to 
perform calculus exitexit}

\item{force, }{boolean if true, previous results in same location inside 
hdf5 will be overwritten.}
}
\value{
hdf5 data file containing a dataset with sd, and mean
}
\description{
This functions gets Standard Deviation (sd) or Mean by Rows or Columns and
store results in hdf5 dataset inside the file
}
\examples{

library(BigDataStatMeth)
# devtools::reload(pkgload::inst("BigDataStatMeth"))
    
# Prepare data and functions
set.seed(123)
Y <- matrix(rnorm(100), 10, 10)
X <- matrix(rnorm(10), 10, 1)
    
# Create hdf5 data file with  data (Y)
bdCreate_hdf5_matrix_file("test.hdf5", Y, "data", "Y", force = TRUE)
bdAdd_hdf5_matrix( X, "test.hdf5",  "data", "X", force = TRUE)

# Get mean and sd        
bdgetSDandMean_hdf5(filename = "test.hdf5", group = "data", dataset = "Y",
                    sd = TRUE, mean = TRUE,byrows = TRUE)
        
}
