% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdgetSDandMean_hdf5}
\alias{bdgetSDandMean_hdf5}
\title{Get sd and Mean by Rows or Columns}
\usage{
bdgetSDandMean_hdf5(
  filename,
  group,
  dataset,
  sd = NULL,
  mean = NULL,
  byrows = NULL,
  wsize = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{filename}{string file name where dataset to normalize is stored}

\item{group}{string Matrix}

\item{dataset}{string Matrix}

\item{sd}{logical (default = TRUE) if TRUE, standard deviation is computed}

\item{mean}{logical (default = TRUE) if TRUE, mean is computed}

\item{byrows}{logical (default = FALSE) if TRUE, sd and mean are computed
by columns, if byrows=TRUE then sd and mean are computed by Rows.}

\item{wsize}{integer (default = 1000), file block size to read to
perform calculus exitexit}

\item{overwrite, }{boolean if true, previous results in same location inside
hdf5 will be overwritten.}
}
\value{
hdf5 data file containing two new datasets, one for sd (if sd is
requested) and another for the mean (if mean is requested). Results are
stored inside a folder mean_sd inside hdf5 data file with names:
sd.\<dataset\>, mean.\<dataset\> respectively
}
\description{
This functions gets Standard Deviation (sd) or Mean by Rows or Columns and
store results in hdf5 dataset inside the file
}
\examples{

library(BigDataStatMeth)
    
# Prepare data and functions
set.seed(123)
Y <- matrix(rnorm(100), 10, 10)
X <- matrix(rnorm(10), 10, 1)
    
# Create hdf5 data file with  data (Y)
bdCreate_hdf5_matrix("test.hdf5", Y, "data", "Y", 
                        overwriteFile = TRUE, 
                        overwriteDataset = FALSE, 
                        unlimited = FALSE)
bdCreate_hdf5_matrix( "test.hdf5", X, "data", "X", 
                       overwriteFile = FALSE, 
                       overwriteDataset = FALSE, 
                       unlimited = FALSE)

# Get mean and sd        
bdgetSDandMean_hdf5(filename = "test.hdf5", group = "data", dataset = "Y",
                    sd = TRUE, mean = TRUE,byrows = TRUE)
                    
# Remove file (used as example)
if (file.exists("test.hdf5")) {
  file.remove("test.hdf5")
}
        
}
