% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdImportTextFile_hdf5}
\alias{bdImportTextFile_hdf5}
\title{Converts text file to hdf5 data file}
\usage{
bdImportTextFile_hdf5(
  filename,
  outputfile,
  outGroup,
  outDataset,
  sep = NULL,
  header = FALSE,
  rownames = FALSE,
  overwrite = FALSE,
  paral = NULL,
  threads = NULL
)
}
\arguments{
\item{filename}{string file name with data to be imported}

\item{outputfile}{file name and path to store imported data}

\item{outGroup}{group name to store the dataset}

\item{outDataset}{dataset name to store the input file in hdf5}

\item{sep}{(optional), by default = "\\t". The field separator string.
Values within each row of x are separated by this string.}

\item{header}{(optional) either a logical value indicating whether the
column names of x are to be written along with x, or a character vector of
column names to be written. See the section on ‘CSV files’ for the meaning
of col.names = NA.}

\item{rownames}{(optional) either a logical value indicating whether the
row names of x are to be written along with x, or a character vector of
row names to be written.}

\item{overwrite}{(optional) either a logical value indicating whether the
output file can be overwritten or not.}

\item{paral, }{(optional, default = TRUE) if paral = TRUE performs parallel
computation else performs seria computation}

\item{threads}{(optional) only if bparal = true, number of concurrent
threads in parallelization if threads is null then threads =  maximum
number of threads available}
}
\value{
none value returned, data are stored in a dataset inside an hdf5 data file.
}
\description{
Converts text file to hdf5 data file
}
