% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdPCA_hdf5}
\alias{bdPCA_hdf5}
\title{PCA Descomposition}
\usage{
bdPCA_hdf5(
  filename,
  group,
  dataset,
  ncomponents = 0L,
  bcenter = FALSE,
  bscale = FALSE,
  k = 2L,
  q = 1L,
  rankthreshold = 0,
  SVDgroup = NULL,
  overwrite = FALSE,
  method = NULL,
  threads = NULL
)
}
\arguments{
\item{filename}{string file name where dataset to normalize is stored}

\item{group}{string specifying the group within the HDF5 file containing
matrix dataset.}

\item{dataset}{string, a string specifying the name of the dataset to
perform calculus.}

\item{ncomponents}{integer, An optional integer specifying the number of
principal components to calculate. Defaults to 0, which computes all components.}

\item{bcenter}{logical, An optional logical indicating whether to center
the data by subtracting the column means. Defaults to FALSE.}

\item{bscale}{logica, An optional logical indicating whether to scale the
data. Defaults to FALSE.}

\item{k}{number of local SVDs to concatenate at each level. Defaults to 2.
This parameter helps optimize the performance and memory usage during PCA
calculations.}

\item{q}{number of levels to compute SVD for PCA.
This parameter helps optimize the performance and memory usage during PCA
calculations.}

\item{rankthreshold}{double, threshold used to determine the range of the array.
The matrix rank is equal to the number of
singular values different from the threshold. By default, threshold = 0 is used
to get the matrix rank , but it can be changed to an approximation of 0.}

\item{SVDgroup}{string. Name of the group where the intermediate SVD results
will be stored or located (If it has been previously calculated).
If falready calculated, this group must contain the d, u and v datasets.}

\item{overwrite}{logical value, if true, the SVD is forced to be computed
although the SVD results exists.}

\item{method}{optional, The method argument can specify different PCA algorithms,
defalut = "auto" possible values are: "auto", "blocks", "full":
\itemize{
\item \code{"auto"}:
The option method = "auto" chooses the "full" or "blocks" method depending on
the size of the matrix to be decomposed
\item \code{"blocks"}:
The PCA can be carried out by blocks, recommended option for large matrices
that do not fit in memory
\item \code{"full"}:
The PCA is performed directly without partitioning the matrix
}}

\item{threads}{integer number of threads used to run PCA}
}
\value{
original file with results in folder PCA/\<datasetname\>
}
\description{
Calculates the Principal Component Analysis (PCA) for datasets stored in HDF5 format.
}
