% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdPCA_hdf5}
\alias{bdPCA_hdf5}
\title{PCA Descomposition}
\usage{
bdPCA_hdf5(
  filename,
  group,
  dataset,
  ncomponents = 0L,
  bcenter = FALSE,
  bscale = FALSE,
  k = 2L,
  q = 1L,
  rankthreshold = 0,
  force = FALSE,
  threads = NULL
)
}
\arguments{
\item{filename}{string, file name where dataset is stored}

\item{group}{string group name  where dataset is stored in file}

\item{dataset}{string dataset name with data to perform PCA}

\item{ncomponents}{integer, number of components to be computed, by default ncomponents = 0, all components are computed}

\item{bcenter}{logical value if true data is centered to zero}

\item{bscale}{logical value, if true data is scaled}

\item{k}{number of local SVDs to concatenate at each level, performance parameter}

\item{q}{number of levels to compute SVD for PCA, performance parameter}

\item{rankthreshold}{double, threshold used to determine the range of the array. The matrix rank is equal to the number of
singular values different from the threshold. By default, threshold = 0 is used to get the matrix rank , but it can be
changed to an approximation of 0.}

\item{force}{logical value, if true, the SVD is forced to be computed although the SVD exists}

\item{threads}{integer number of threads used to run PCA}
}
\value{
original file with results in folder PCA/<datasetname>
}
\description{
Compute PCA
}
