% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdPCA_hdf5}
\alias{bdPCA_hdf5}
\title{PCA Descomposition}
\usage{
bdPCA_hdf5(filename, group, dataset, threads)
}
\arguments{
\item{filename}{string, file name where dataset is stored}

\item{group}{string group name  where dataset is stored in file}

\item{dataset}{string dataset name with data to perform PCA}
}
\value{
original file with results in folder PCA/<datasetname>
}
\description{
Compute PCA
}
