% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdPCA_hdf5}
\alias{bdPCA_hdf5}
\title{PCA Descomposition with HDF5 files}
\usage{
bdPCA_hdf5(
  filename,
  group,
  dataset,
  ncomponents = 0L,
  bcenter = FALSE,
  bscale = FALSE,
  k = 2L,
  q = 1L,
  rankthreshold = 0,
  SVDgroup = NULL,
  overwrite = FALSE,
  method = NULL,
  threads = NULL
)
}
\arguments{
\item{group}{string specifying the group within the HDF5 file containing
matrix dataset.}

\item{dataset}{string, a string specifying the name of the dataset to
perform calculus.}

\item{ncomponents}{integer, An optional integer specifying the number of
principal components to calculate. Defaults to 0, which computes all components.}

\item{bcenter}{logical (optional). If TRUE, the data is centered by
subtracting the column means (ignoring NAs) of \code{x} from their corresponding columns.
If FALSE (default), no centering is performed.}

\item{bscale}{(optional). If TRUE, the data is scaled by dividing
the (centered) columns of \code{x} by their standard deviations if \code{bcenter}
is TRUE, or by the root mean square otherwise. If FALSE (default), no
scaling is performed.}

\item{k}{numerical, number of local SVDs to concatenate at each level.
Defaults is set to 2.
This parameter helps optimize the performance and memory usage during PCA
calculations.}

\item{q}{numerical, number of levels to compute SVD for PCA.
This parameter helps optimize the performance and memory usage during PCA
calculations.}

\item{rankthreshold}{\code{double}. Threshold used to determine the range of
the matrix. The matrix rank is defined as the number of singular values that
differ from the threshold. By default, \code{threshold = 0} is used to compute
the matrix rank, but it can be adjusted to a value close to zero for
approximations.}

\item{SVDgroup}{string. Name of the group where the intermediate SVD results
will be stored or located (If it has been previously calculated).
If falready calculated, this group must contain the d, u and v datasets.}

\item{overwrite}{logical value, If TRUE, forces the recalculation of results
even if they already exist.}

\item{method}{optional, The method argument can specify different PCA algorithms,
defalut = "auto" possible values are: "auto", "blocks", "full":
\itemize{
\item \code{"auto"}:
The option method = "auto" chooses the "full" or "blocks" method depending on
the size of the matrix to be decomposed
\item \code{"blocks"}:
The PCA can be carried out by blocks, recommended option for large matrices
that do not fit in memory
\item \code{"full"}:
The PCA is performed directly without partitioning the matrix
}}

\item{threads}{integer (optional), an optional parameter specifying the
number of threads to use.}
}
\value{
original file with results in folder PCA/\<datasetname\>
}
\description{
Calculates the Principal Component Analysis (PCA) for datasets stored in HDF5 format.
}
