% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdpseudoinv}
\alias{bdpseudoinv}
\title{Pseudo-Inverse}
\usage{
bdpseudoinv(X, threads = NULL)
}
\arguments{
\item{X}{Singular matrix (m x n)}

\item{threads}{(optional) only if bparal = true, number of concurrent
threads in parallelization if threads is null then threads =  maximum number of threads available}
}
\value{
Pseudo-inverse matrix of A
}
\description{
Compute the pseudo-inverse of a singular matrix
}
