% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdpseudoinv}
\alias{bdpseudoinv}
\title{Pseudo-Inverse}
\usage{
bdpseudoinv(X)
}
\arguments{
\item{Singular}{matrix (m x n)}
}
\value{
Pseudo-inverse matrix of A
}
\description{
Compute the pseudo-inverse of a singular matrix
}
