% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdpseudoinv_hdf5}
\alias{bdpseudoinv_hdf5}
\title{Pseudo-Inverse}
\usage{
bdpseudoinv_hdf5(
  filename,
  group,
  dataset,
  outgroup = NULL,
  outdataset = NULL,
  overwrite = NULL,
  threads = NULL
)
}
\arguments{
\item{filename, }{character array with the name of an existin hdf5 data file containing the dataset}

\item{group, }{character array indicating the input group where the dataset is stored}

\item{dataset, }{character array indicating the input dataset that contains an m x n singular matrix}

\item{outgroup}{optional, character array with output group name where we want to
store results if not provided then results are stored in the same group as original dataset}

\item{outdataset}{character array with output dataset name where we want to store results}

\item{overwrite, }{optional boolean if true, previous results in same location inside
hdf5 will be overwritten, by default force = false, data was not overwritten.}

\item{threads}{optional parameter. Integer with numbers of threads to be used}
}
\value{
Pseudo-inverse matrix of A
}
\description{
Compute the pseudo-inverse of a singular matrix in hdf5 data files
}
