% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdQR}
\alias{bdQR}
\title{QR Decomposition}
\usage{
bdQR(X, thin = NULL)
}
\arguments{
\item{X}{a real square matrix}

\item{boolean}{thin, if thin = true returns Q thin  decomposition else returns Q full decomposition, default thin = false}
}
\value{
List with orthogonal matrix \code{Q}  and upper triangular matrix \code{R}
}
\description{
This function compute QR decomposition (also called a QR factorization) 
of a matrix \code{A} into a product \code{A = QR} of an 
orthogonal matrix Q and an upper triangular matrix R.
}
