% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdReduce_hdf5_dataset}
\alias{bdReduce_hdf5_dataset}
\title{Reduce hdf5 dataset}
\usage{
bdReduce_hdf5_dataset(
  filename,
  group,
  reducefunction,
  outgroup = NULL,
  outdataset = NULL,
  overwrite = FALSE,
  remove = FALSE
)
}
\arguments{
\item{filename, }{character array with the name of the file where datasets are stored}

\item{group, }{character array with the input group name where the data sets are stored}

\item{reducefunction, }{single character with function to apply, can be '+' or  '-'}

\item{outgroup, }{optional character array with the group name where the dataset will be stored after reduction. If `outgroup` is NULL, the resulting dataset is stored in the same input group.}

\item{outdataset, }{optional character with the dataset name for the resulting dataset after reduction if `outdataset` is NULL, then the input group name is used as outdataset name}

\item{overwrite, }{boolean if true, previous results in same location inside hdf5 will be overwritten.}

\item{remove, }{boolean if true, removes original matrices, by default bremove = false.}
}
\value{
Full matrix with results from reduction
}
\description{
Reduce hdf5 datasets inside a group by rows or columns and store complete matrix inside the hdf5 data file.
}
