% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdRemovelowdata_hdf5}
\alias{bdRemovelowdata_hdf5}
\title{Remove SNPs in hdf5 omic dataset with low data}
\usage{
bdRemovelowdata_hdf5(
  filename,
  group,
  dataset,
  outgroup,
  outdataset,
  pcent,
  bycols,
  overwrite = NULL
)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{group, }{character array indicating the input group where the data set to be imputed is.}

\item{dataset, }{character array indicating the input dataset to be imputed}

\item{outgroup, }{character array indicating group where the data set will be 
saved after remove data with if `outgroup` is NULL, output dataset is stored 
in the same input group.}

\item{outdataset, }{character array indicating dataset to store the resulting 
data after imputation if `outdataset` is NULL, input dataset will be overwritten.}

\item{pcent, }{by default pcent = 0.5. Numeric indicating the percentage to be 
considered to remove SNPs, SNPS with percentage equal or higest will be removed from data}

\item{bycols, }{boolean by default = true, if true, indicates that SNPs are in 
cols, if SNPincols = false indicates that SNPs are in rows.}

\item{overwrite, }{optional boolean if true, previous results in same location 
inside hdf5 will be overwritten, by default overwrite = false, data was not overwritten.}
}
\value{
Original hdf5 data file with imputed data
}
\description{
Remove SNPs in hdf5 omic dataset with low data
}
