% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdSVD_hdf5}
\alias{bdSVD_hdf5}
\title{Block SVD decomposition with HDF5 files}
\usage{
bdSVD_hdf5(
  file,
  group = NULL,
  dataset = NULL,
  k = 2L,
  q = 1L,
  bcenter = TRUE,
  bscale = TRUE,
  rankthreshold = 0,
  overwrite = NULL,
  method = NULL,
  threads = NULL
)
}
\arguments{
\item{group}{string specifying the group within the HDF5 file containing
matrix dataset.}

\item{dataset}{string, a string specifying the name of the dataset to
perform calculus.}

\item{k}{number of local SVDs to concatenate at each level. Defaults to 2.
This parameter helps optimize the performance and memory usage during PCA
calculations.}

\item{q}{number of levels to compute SVD for PCA.
This parameter helps optimize the performance and memory usage during PCA
calculations.}

\item{bcenter}{logical (optional). If TRUE (default), the data is centered
by subtracting the column means (ignoring NAs) of the \code{dataset} from their
corresponding columns. If FALSE, no centering is performed.}

\item{bscale}{(optional). If TRUE (default), the data is scaled by dividing
the (centered) columns of \code{x} by their standard deviations if \code{bcenter}
is TRUE, or by the root mean square otherwise. If FALSE, no scaling is
performed.}

\item{rankthreshold}{\code{double}. Threshold used to determine the range of
the matrix. The matrix rank is defined as the number of singular values that
differ from the threshold. By default, \code{threshold = 0} is used to compute
the matrix rank, but it can be adjusted to a value close to zero for
approximations.}

\item{overwrite}{logical value, If TRUE, forces the recalculation of results
even if they already exist.}

\item{method}{optional, defalut is "auto" possible values are: "auto",
"blocks", "full":
* \code{"auto"}:
The option method = "auto" chooses the "full" or
"blocks" method depending on the size of the matrix to be decomposed
* \code{"blocks"}:
The SVD decomposition can be carried out by blocks, recommended option
for large matrices that do not fit in memory
* \code{"full"}:
The SVD decomposition is performed directly without partitioning the matrix}

\item{threads}{integer (optional), an optional parameter specifying the
number of threads to use.}
}
\value{
three dataset inside HDF5 data files with the singular values and
left and right singular vectors of the dataset:
\itemize{
\item \code{"u"}:
eigenvectors of AA^t, mxn and column orthogonal matrix
\item \verb{"v}:
eigenvectors of A^tA, nxn orthogonal matrix
\item \code{"d"}:
singular values, nxn diagonal matrix (non-negative real values)
}
}
\description{
This function computes the singular values and left singular vectors of a
real nxp matrix using Block SVD decomposition with an incremental algorithm.
The input matrix is read from an HDF5 file, and the results are saved in the
same file.
}
