% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdtCrossprod_hdf5}
\alias{bdtCrossprod_hdf5}
\title{tCrossprod with hdf5 matrix}
\usage{
bdtCrossprod_hdf5(
  filename,
  group,
  A,
  B = NULL,
  groupB = NULL,
  block_size = NULL,
  mixblock_size = NULL,
  paral = NULL,
  threads = NULL,
  outgroup = NULL,
  outdataset = NULL,
  overwrite = NULL
)
}
\arguments{
\item{filename}{string specifying the path to the HDF5 file}

\item{group}{string specifying the group within the HDF5 file containing
matrix A.}

\item{A}{string specifying the dataset name for matrix A.
the data matrix to be used in calculus}

\item{B}{string specifying the dataset name for matrix B.}

\item{groupB}{string, (optional), An optional string specifying the group
for matrix B. Defaults to the value of \code{group} if not provided.}

\item{block_size}{integer (optional), an optional parameter specifying the
block size for processing the matrices. If not provided, a default block
size is used. The block size should be chosen based on the available memory
and the size of the matrices}

\item{mixblock_size}{only for debug pourpose}

\item{paral}{boolean (optional), an optional parameter to enable parallel
computation. Defaults to FALSE. Set \code{paral = true} to force parallel execution}

\item{threads}{integer (optional), an optional parameter specifying the
number of threads to use if paral = TRUE. Ignored if paral = FALSE.}

\item{outgroup}{string (optional), An optional parameger specifying the group
where the output matrix will be stored. If NULL, the output will be stored
in the default group "OUTPUT".}

\item{outdataset}{string (optional), An optional parameter specifying the
dataset name for the output matrix. If NULL, the default name will be
constructed as "tCrossProd_" concatenated with the name of dataset A
"\emph{x}" and the name of dataset B.}

\item{overwrite}{logical (optional), An optional parameter to indicate whether
existing results in the HDF5 file should be overwritten. Defaults to FALSE.
If FALSE and the dataset already exists, an error will be displayed, and
no calculations will be performed. If TRUE and a dataset with the same
name as specified in outdataset already exists, it will be overwritten.}
}
\value{
no value
}
\description{
This function performs the transposed cross product of one or two matrices
}
\details{
For a single matrix \eqn{A}, the cross product is defined as \eqn{A A^t},
where \eqn{A^t} is the transpose of \eqn{A}. For two matrices \eqn{A} and
\eqn{B}, the cross product is \eqn{A B^t}. This operation is often used in
linear algebra for projections and other computations.
}
\examples{
  
  library(BigDataStatMeth)
  library(rhdf5)
  
  N = 1000
  M = 1000
  
  set.seed(555)
  a <- matrix( rnorm( N*M, mean=0, sd=1), N, M) 
  
  bdCreate_hdf5_matrix( filename = "test_temp.hdf5", 
                        object = a, group = "INPUT", 
                        dataset = "datasetA",
                        transp = FALSE,
                        overwriteFile = TRUE, 
                        overwriteDataset = FALSE, 
                        unlimited = FALSE)
                        
    file <- "test_temp.hdf5"
    dataset <- "results/res"
    
    bdtCrossprod_hdf5( filename = "test_temp.hdf5", group = "INPUT", 
                       A = "datasetA", outgroup = "results", 
                       outdataset = "res", overwrite = TRUE ) # 
                       
    # Check results
    resr <- tcrossprod(a)
    res <-  h5read(file,dataset)
    all.equal( resr, res)
    
    bdtCrossprod_hdf5(filename = "test_temp.hdf5", group = "INPUT", 
                       A = "datasetA", outgroup = "results", 
                       outdataset = "res", block_size = 1024, 
                       overwrite = TRUE ) # 
    
    # Check results
    resr <- tcrossprod(a)
    res <-  h5read(file,dataset)
    all.equal( resr, res)
  
    # Remove file (used as example)
    if (file.exists("test_temp.hdf5")) {
      file.remove("test_temp.hdf5")
    }
  

}
