% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdwproduct}
\alias{bdwproduct}
\title{Matrix - Weighted vector Multiplication with numerical or DelayedArray data}
\usage{
bdwproduct(a, w, op)
}
\arguments{
\item{w}{vector with weights}

\item{op}{string indicating if operation  "Xw" or "wX"}

\item{X}{numerical or Delayed Array matrix}

\item{bparal}{(optional, default = true) if bparal=true performs parallel computation.}
}
\value{
numerical matrix
}
\description{
This function performs a weighted product of a matrix(X) with a weighted diagonal matrix (w)
}
\examples{
n <- 100
p <- 60

X <- matrix(rnorm(n*p), nrow=n, ncol=p)

u <- runif(n)
w <- u * (1 - u)
bdwproduct(X, w,"xtwx")
bdwproduct(X, w,"xwxt")

# with Delayed Array

DX <- DelayedArray(X)
dw <- DelayedArray(w)

bdwproduct(DX, dw,"xtwx")
bdwproduct(DX, dw,"xwxt")

}
