% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdWriteDiagonal_hdf5}
\alias{bdWriteDiagonal_hdf5}
\title{Write diagonal matrix}
\usage{
bdWriteDiagonal_hdf5(diagonal, filename, group, dataset)
}
\arguments{
\item{diagonal, }{numeric vector with diagonal elements to be written in existing dataset.}

\item{filename, }{character array with the name of an existin hdf5 data file containing the dataset to be modified}

\item{group, }{character array indicating the input group where the data set to be modified.}

\item{dataset, }{character array indicating the input dataset to be modified}
}
\value{
Original hdf5 dataset with new diagonal elements
}
\description{
Write diagonal matrix to an existing dataset inside hdf5
}
\examples{
library(BigDataStatMeth)
library(rhdf5)

# Prepare data and functions
X <- matrix(rnorm(100), 10, 10)
diagonal <- c(1,2,3,4,5,6,7, 8, 9, 10)

# Create hdf5 data file with  data (Y)
bdCreate_hdf5_matrix_file("test_file.hdf5", X, "data", "X", force = TRUE)

# Update diagonal
bdWriteDiagonal_hdf5(diagonal, "test_file.hdf5", "data", "X")

}
