% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdWriteDimnames_hdf5}
\alias{bdWriteDimnames_hdf5}
\title{Exists hdf5 element}
\usage{
bdWriteDimnames_hdf5(filename, group, dataset, rownames, colnames)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{group, }{string with name of the group where the new dataset will be 
created}

\item{dataset, }{string with name for the new dataset}

\item{rownames, }{character vector, with the rownames, if rownames is NULL 
no rownames are written to the dataset}

\item{colnames, }{character vector, with the colnames, if colnames is NULL
no colnames are written to the dataset}
}
\value{
boolean, true if element exists in hdf5 data faile or false if not.
}
\description{
Query if exists element inside hdf5 data file
}
\examples{

# Prepare data to write dataset inside a file and test if exists

}
