% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdWriteOppsiteTriangularMatrix_hdf5}
\alias{bdWriteOppsiteTriangularMatrix_hdf5}
\title{Write Upper/Lower triangular matrix}
\usage{
bdWriteOppsiteTriangularMatrix_hdf5(
  filename,
  group,
  dataset,
  copytolower = NULL,
  elementsBlock = 1000000L
)
}
\arguments{
\item{filename, }{character array with the name of an existin hdf5 data file containing the dataset to be modified}

\item{group, }{character array indicating the input group where the data set to be modified.}

\item{dataset, }{character array indicating the input dataset to be modified}

\item{copytolower, }{boolean with default value = false. If true, sets lower 
triangular matrix using upper triangular matrix, if lower=false (default 
value) sets upper triangular matrix using lower triangular matrix.}

\item{elementsBlock, }{optional integer defines de maximum number of elements 
to read from hdf5 data file in each block. By default this value is set 
to 10000. If matrix is bigger thant 5000x5000 then block is set to number 
of rows or columns x 2}
}
\description{
Write diagonal matrix to an existing dataset inside hdf5
}
\examples{
library(BigDataStatMeth)

# Prepare data and functions
X <- matrix(rnorm(100), 10, 10)
X.1 <- X
X[lower.tri(X)] <- 0
# Create hdf5 data file with  data (Y)
bdCreate_hdf5_matrix("test_file.hdf5", X, "data", "X", force = TRUE)
# Update Lower triangular matrix in hdf5
bdWriteOppsiteTriangularMatrix_hdf5(filename = "test_file.hdf5", 
        group = "data", dataset = "X", copytolower = TRUE, elementsBlock = 10)

X <- X.1
X[upper.tri(X)] <- 0
# CAdd matrix data to a file
bdCreate_hdf5_matrix( "test_file.hdf5", X, "data", "Y", force = TRUE)
# Update Upper triangular matrix in hdf5
bdWriteOppsiteTriangularMatrix_hdf5(filename = "test_file.hdf5", 
        group = "data", dataset = "Y", copytolower = FALSE, elementsBlock = 10)

}
