% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdwXw}
\alias{bdwXw}
\title{Matrix - Weighted vector Multiplication with numerical or DelayedArray data}
\usage{
bdwXw(X, W, op)
}
\arguments{
\item{X}{numerical or Delayed Array matrix}

\item{op}{string indicating if operation  "Xw" , "wX" or "wXw"}

\item{w}{vector with weights}
}
\value{
numerical matrix
}
\description{
This function performs a weighted product of a matrix(X) with a weighted diagonal matrix (w)
}
\examples{
n <- 100
p <- 60

X <- matrix(rnorm(n*p), nrow=n, ncol=p)

u <- runif(n)
w <- u * (1 - u)

# by columnes
bdwXw(X, w,"wX") 

# by rows
bdwXw(X, w,"Xw")

# by columns and rows
bdwXw(X, w,"wXw")

# with Delayed Array

DX <- DelayedArray(X)
dw <- DelayedArray(w)

bdwXw(DX, dw,"wX")
bdwXw(DX, dw,"Xw")

}
