% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/blockmult_hdf5.R
\name{blockmult_hdf5}
\alias{blockmult_hdf5}
\title{Multiply hdf5 matrix}
\usage{
blockmult_hdf5(
  filename,
  group,
  a,
  b,
  block_size = 128,
  paral = FALSE,
  threads = NULL,
  mixblock_size = 128,
  outgroup = "OUTPUT"
)

blockmult_hdf5(
  filename,
  group,
  a,
  b,
  block_size = 128,
  paral = FALSE,
  threads = NULL,
  mixblock_size = 128,
  outgroup = "OUTPUT"
)
}
\arguments{
\item{filename}{string file name where dataset to normalize is stored}

\item{group}{string with the group name where matrix is stored inside HDF5 file}

\item{a}{a double matrix.}

\item{b}{a double matrix.}

\item{block_size}{(optional, defalut = 128) block size to make matrix multiplication, if `block_size = 1` no block size is applied (size 1 = 1 element per block)}

\item{paral, }{(optional, default = FALSE) paral = true --> TO BE IMPLEMENTED}

\item{threads}{(optional) only if bparal = true, number of concurrent threads in parallelization if threads is null then threads =  maximum number of threads available}

\item{mixblock_size}{(optiona, default = 128), only if we are working with big matrix and parallel computation = true. 
Block size for mixed computation in big matrix parallel. Size of the block to be used to perform parallelized memory 
memory of the block read from the disk being processed.}

\item{outgroup}{(optional) string with group name where we want to store the result matrix}

\item{dataset}{string or Delayed Array Matrix}

\item{bcenter}{logical (default = TRUE) if TRUE, centering is done by subtracting the column means}

\item{bscale}{logical (default = TRUE) if TRUE, centering is done by subtracting the column means}

\item{wsize}{integer (default = 1000), file block size to read to perform normalization}
}
\value{
file with scaled, centered or scaled and centered dataset

A list with an HDF5 object with numerical matrix and  HDF5 file name with results
\itemize{
 #### \item{"res"}{rhdf5 object with result matrix - link to hdf5 file contents . IMPORTANT !!, we have to close the object after fihish to work}
 \item{"file"}{string with hdf5 file name with result and original matrixs}
 \item{"dataset"}{string complete path inside hdf5 file where results are stored.}
}
}
\description{
This function multiply matrix stored in hdf5 data file

This function performs a block matrix-matrix multiplication with mattrix stored in HDF5 file
}
\examples{
  a = "See vignette"
# with numeric matrix

library(DelayedArray)

m <- 500
k <- 1500
n <- 400
A <- matrix(rnorm(n*k), nrow=n, ncol=k)
B <- matrix(rnorm(n*k), nrow=k, ncol=n)

blockmult(A, B, 128, TRUE)

# with Delaeyd Array
AD <- DelayedArray(A)
BD <- DelayedArray(B)

blockmult_hdf5( AD, BD, 128, TRUE)

}
