% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{colesterol}
\alias{colesterol}
\title{Dataset colesterol}
\description{
This is a dataset containing simulated variables related to colesterol
}
\section{colesterol.csv}{


\describe{
  \item{TCholesterol}{TCholesterol}
  \item{Age}{Age}
  \item{Insulin}{Insulin}
  \item{Creatinine}{Creatinine}
  \item{BUN }{BUN}
  \item{LLDR}{LLDR}
  \item{Triglycerides}{Triglycerides}
  \item{HDL_C}{HDL_C}
  \item{LDL_C}{LDL_C}
  \item{Sex}{Sex}
}

This data is used in bdImport_text_to_hdf5().
}

