% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Create_HDF5_matrix_file}
\alias{Create_HDF5_matrix_file}
\title{Create hdf5 data file and write data to it}
\usage{
Create_HDF5_matrix_file(
  filename,
  object,
  group = NULL,
  dataset = NULL,
  transp = NULL
)
}
\arguments{
\item{filename, }{character array indicating the name of the file to create}

\item{object}{numerical data matrix}

\item{group, }{character array indicating folder name to put the matrix in hdf5 file}

\item{dataset, }{character array indicating the dataset name to store the matix data}

\item{transp}{boolean, if trans=true matrix is stored transposed in hdf5 file}
}
\value{
none
}
\description{
Creates a hdf5 file with numerical data matrix,
}
