% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_core.R
\name{dim.HDF5Matrix}
\alias{dim.HDF5Matrix}
\title{Dimensions of an HDF5Matrix}
\usage{
\method{dim}{HDF5Matrix}(x)
}
\arguments{
\item{x}{An \code{HDF5Matrix} object}
}
\value{
Integer vector \code{c(nrows, ncols)}
}
\description{
Dimensions of an HDF5Matrix
}
\examples{
\donttest{
tmp <- tempfile(fileext = ".h5")
rhdf5::h5createFile(tmp)
rhdf5::h5createGroup(tmp, "data")
rhdf5::h5write(matrix(1:100, 10, 10), tmp, "data/matrix")
X <- hdf5_matrix(tmp, "data/matrix")

dim(X)    # c(10, 10)
nrow(X)   # 10
ncol(X)   # 10

X$close()
unlink(tmp)
}

}
