% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bdCCA.R
\name{getQRbyBlocks}
\alias{getQRbyBlocks}
\title{QR by blocks}
\usage{
getQRbyBlocks(strdataset, file, mblocks, center, scale, bcols, overwrt)
}
\arguments{
\item{strdataset}{string, dataset path within the hdf5 data file from which we want to calculate the QR}

\item{mblocks}{number of blocks in which we want to partition the matrix to perform the calculations}

\item{center, }{boolean, if true, dataset is centered to perform calculus}

\item{scale, }{boolean, if true, dataset is centered to perform calculus}

\item{overwrt, }{boolean, if true, datasets existing inside a file must be overwritten if we are using the same names}

\item{filename}{string file name where dataset to normalize is stored}
}
\value{
hdf5 data file with CCA results,
}
\description{
This function is an application of the BigDataStatMeth functions to generate new methods. This function perform a QR
from two matrices stored in hdf5 data file. This function applies matrix partitioning, merge bloks to create a full matrix, apply a function to different blocks...
}
\examples{

}
