% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_core.R
\name{is_open}
\alias{is_open}
\title{Check if HDF5Matrix is open}
\usage{
is_open(x)
}
\arguments{
\item{x}{An \code{HDF5Matrix} object}
}
\value{
Logical. \code{TRUE} if object is valid and open, \code{FALSE} otherwise.
}
\description{
Check whether an HDF5Matrix object is still valid and open.
}
\examples{
\donttest{
tmp <- tempfile(fileext = ".h5")

X  <- hdf5_create_matrix(tmp, "data/matrix", data = matrix(rnorm(100), 10, 10))

X <- hdf5_matrix(tmp, "data/matrix")
is_open(X)  # TRUE

close(X)
is_open(X)  # FALSE

unlink(tmp)
}

}
