% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LOOE}
\alias{LOOE}
\title{LOOE}
\usage{
LOOE(X, Y, paral, nl = NULL, ml = NULL, l = NULL)
}
\arguments{
\item{X}{...}

\item{Y}{...}

\item{paral}{...}

\item{nl}{...}

\item{ml}{...}

\item{l}{...}
}
\value{
coef ...

Ginv  ...

lambda.min  ...

lambdas  ...

looe ...
}
\description{
This function performs ...
}
\examples{
n <- 500
p <- 200
M <- matrix(rnorm(n*p), nrow=n, ncol=p)
Y <- 2.4*M[,1] + 1.6*M[,2] - 0.4*M[,5]

MD <- DelayedArray(M)
YD <- DelayedArray(as.matrix(Y))

looe <- LOOE(M,Y,paral=TRUE),
looe_delayed <- LOOE(MD,YD,paral=TRUE)

looe$lambda.min
looe$lambdas
looe_delayed$coef[1:10]
}
