% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Normalize_Data}
\alias{Normalize_Data}
\title{Normalize Delayed Array matrix}
\usage{
Normalize_Data(x, bcenter = NULL, bscale = NULL)
}
\arguments{
\item{bcenter}{logical (default = TRUE) if TRUE, centering is done by subtracting the column means}

\item{bscale}{logical (default = TRUE) if TRUE, centering is done by subtracting the column means}

\item{X}{numerical or Delayed Array Matrix}
}
\value{
numerical matrix
}
\description{
This function performs a numerical or Delayed Array matrix normalization
}
\examples{
m <- 500
n <- 100 
x <- matrix(rnorm(m*n), nrow=m, ncol=n)

# with numeric matrix
Normalize_Data(x)

# with Delaeyd Array
Dx <- DelayedArray(x)

# Center and scale
Normalize_Data(Dx)

# Only scale
Normalize_Data(Dx, bcenter = FALSE)

# Only center
Normalize_Data(Dx, bscale = FALSE)

}
