% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parXYProdBlock}
\alias{parXYProdBlock}
\title{Block matrix multiplication with Delayed Array Object}
\usage{
parXYProdBlock(X, Y, op = NULL)
}
\arguments{
\item{op, }{(optional, default = "xy"), if op="xy" then performs the x\%*\%y matrix multiplication, if op = "xty" preforms t(X)\%*\% Y, if op = "xyt" performs X\%*\%t(Y)}

\item{a}{a double matrix.}

\item{b}{a double matrix.}
}
\value{
numerical matrix
}
\description{
This function performs a block matrix-matrix multiplication with numeric matrix or Delayed Arrays
}
\examples{

library(DelayedArray)

# with numeric matrix
m <- 500
k <- 1500
n <- 400
A <- matrix(rnorm(n*p), nrow=n, ncol=k)
B <- matrix(rnorm(n*p), nrow=k, ncol=n)

Bblockmult(A,B,128, TRUE)

# with Delaeyd Array
AD <- DelayedArray(A)
BD <- DelayedArray(B)

Bblockmult(AD,BD,128, TRUE)

}
